moved CarbonEmissionQueryException to adapter package
