+++ carbon-reductor-core/src/main/java/de/envite/greenbpm/carbonreductor/core/adapter/watttime/CarbonAwareSdkClient.java
@@ -3,7 +3,7 @@ package de.envite.greenbpm.carbonreductor.core.adapter.watttime;
 import de.envite.greenbpm.api.carbonawaresdk.ApiResponse;
 import de.envite.greenbpm.api.carbonawaresdk.api.CarbonAwareApi;
 import de.envite.greenbpm.api.carbonawaresdk.model.EmissionsForecastDTO;
-import de.envite.greenbpm.carbonreductor.core.adapter.watttime.exception.CarbonEmissionQueryException;
+import de.envite.greenbpm.carbonreductor.core.adapter.exception.CarbonEmissionQueryException;
 import de.envite.greenbpm.carbonreductor.core.domain.model.EmissionTimeframe;
 import de.envite.greenbpm.carbonreductor.core.domain.model.input.Timeshift;
 import de.envite.greenbpm.carbonreductor.core.domain.model.input.location.Location;
