+++ carbon-reductor-core/src/main/java/de/envite/greenbpm/carbonreductor/core/domain/service/DelayCalculatorService.java
@@ -1,6 +1,6 @@
 package de.envite.greenbpm.carbonreductor.core.domain.service;
 
-import de.envite.greenbpm.carbonreductor.core.adapter.watttime.exception.CarbonEmissionQueryException;
+import de.envite.greenbpm.carbonreductor.core.adapter.exception.CarbonEmissionQueryException;
 import de.envite.greenbpm.carbonreductor.core.domain.model.CarbonReduction;
 import de.envite.greenbpm.carbonreductor.core.domain.model.input.Timeshift;
 import de.envite.greenbpm.carbonreductor.core.usecase.out.CarbonEmissionQuery;
