+++ carbon-reductor-core/src/test/java/de/envite/greenbpm/carbonreductor/core/adapter/carbonawarecomputing/CarbonAwareComputingApiClientTest.java
@@ -3,7 +3,7 @@ package de.envite.greenbpm.carbonreductor.core.adapter.carbonawarecomputing;
 import de.envite.greenbpm.api.carbonawarecomputing.api.ForecastApi;
 import de.envite.greenbpm.api.carbonawarecomputing.model.EmissionsData;
 import de.envite.greenbpm.api.carbonawarecomputing.model.EmissionsForecast;
-import de.envite.greenbpm.carbonreductor.core.adapter.watttime.exception.CarbonEmissionQueryException;
+import de.envite.greenbpm.carbonreductor.core.adapter.exception.CarbonEmissionQueryException;
 import de.envite.greenbpm.carbonreductor.core.domain.model.EmissionTimeframe;
 import de.envite.greenbpm.carbonreductor.core.domain.model.emissionframe.ForecastedValue;
 import de.envite.greenbpm.carbonreductor.core.domain.model.emissionframe.OptimalTime;
