+++ carbon-reductor-core/src/test/java/de/envite/greenbpm/carbonreductor/core/domain/service/DelayCalculatorServiceTest.java
@@ -1,7 +1,7 @@
 package de.envite.greenbpm.carbonreductor.core.domain.service;
 
 import de.envite.greenbpm.carbonreductor.core.TestDataGenerator;
-import de.envite.greenbpm.carbonreductor.core.adapter.watttime.exception.CarbonEmissionQueryException;
+import de.envite.greenbpm.carbonreductor.core.adapter.exception.CarbonEmissionQueryException;
 import de.envite.greenbpm.carbonreductor.core.domain.model.CarbonReduction;
 import de.envite.greenbpm.carbonreductor.core.domain.model.emissionframe.ForecastedValue;
 import de.envite.greenbpm.carbonreductor.core.domain.model.emissionframe.OptimalTime;
